requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.1
awsebcli==3.18.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
blurhash==1.1.4
botocore==1.15.46
bs4==0.0.1
cachetools==4.1.0
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
codegen==1.0
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
Cython==0.29.15
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.41.0
django-compat==1.0.15
django-filter==2.2.0
django-mssql==1.8
django-mssql-backend==2.5.0
django-pyodbc-azure==2.1.0.0
docutils==0.15.2
Faker==3.0.0
feedparser==5.1.3
Flask==1.1.2
future==0.16.0
gcloud==0.18.3
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
googlevoice==3.0.0
grpcio==1.28.1
gunicorn==20.0.4
html5lib==0.999
httplib2==0.17.0
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.15.1
jws==0.1.3
lxml==4.5.0
MarkupSafe==1.1.1
Mastodon.py==1.5.1
mysql-connector==2.2.9
natsort==3.4.0
nltk==3.5
numpy==1.18.2
oauth2client==4.1.2
oauthlib==3.1.0
opencv-python==4.2.0.34
pandas==1.0.3
parsel==1.6.0
pathspec==0.5.9
patterns==0.3
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycode-similar==1.3
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.9.7
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyodbc==4.0.28
pypiwin32==223
PyQt5==5.14.1
PyQt5-sip==12.7.1
PyQt5Designer==5.14.1
Pyrebase4==4.3.0
python-dateutil==2.8.0
python-jwt==2.0.1
python-magic==0.4.15
python3-openid==3.1.0
pytz==2019.3
pywin32==227
PyYAML==5.3.1
razorpay==1.2.0
regex==2020.6.8
requests==2.3.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
selenium==3.141.0
semantic-version==2.5.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.16
sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.3
tqdm==4.46.1
urllib3==1.15
utils==1.0.1
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
w3lib==1.22.0
wcwidth==0.1.9
Werkzeug==1.0.1
whitenoise==5.0.1
zope.interface==5.1.0
